+++ src/components/Charts/PriceChart/utils.ts
