+++ src/components/Tokens/TokenDetails/PriceChart.tsx
@@ -7,7 +7,7 @@ import { Line } from '@visx/shape'
 import AnimatedInLineChart from 'components/Charts/AnimatedInLineChart'
 import FadedInLineChart from 'components/Charts/FadeInLineChart'
 import { getTimestampFormatter, TimestampFormatterType } from 'components/Charts/PriceChart/format'
-import { getTicks } from 'components/Charts/PriceChart/util'
+import { getTicks } from 'components/Charts/PriceChart/utils'
 import { ArrowChangeDown } from 'components/Icons/ArrowChangeDown'
 import { ArrowChangeUp } from 'components/Icons/ArrowChangeUp'
 import { MouseoverTooltip } from 'components/Tooltip'
